moved fetch into a hook
